moved multi-d into main class
