+++ src/index.ts
@@ -1,7 +1,6 @@
 import GenerateArray from './GenerateArray.ts';
-import Multidimensional from "./MultidimensionalArray.ts";
 import CustomArray from "./CustomArrayGen.ts";
 import ArrayUtils from "./ArrayUtils.ts";
 import GenerateArrayError from "./GenerateArraysError.ts";
 
-export { GenerateArray, Multidimensional, CustomArray, ArrayUtils, GenerateArrayError };
+export { GenerateArray, CustomArray, ArrayUtils, GenerateArrayError };
